refactor(exporters): move json schema traverser logic to utils
