+++ src/s2dm/exporters/utils.py
@@ -4,16 +4,26 @@ from dataclasses import dataclass
 from enum import Enum
 from itertools import product
 from pathlib import Path
-from typing import Any
+from typing import Any, cast
 
 from ariadne import load_schema_from_path
 from graphql import (
     GraphQLEnumType,
+    GraphQLInputObjectType,
+    GraphQLInterfaceType,
+    GraphQLList,
+    GraphQLNonNull,
     GraphQLString,
+    GraphQLType,
+    GraphQLUnionType,
     build_schema,
     get_named_type,
+    is_input_object_type,
+    is_interface_type,
     is_list_type,
     is_non_null_type,
+    is_object_type,
+    is_union_type,
 )
 from graphql.type import (
     GraphQLField,
@@ -528,3 +538,81 @@ def search_schema(
                 continue
 
     return results
+
+
+def get_referenced_types(graphql_schema: GraphQLSchema, root_type: str) -> set[GraphQLType]:
+    """
+    Find all GraphQL types referenced from the root type through graph traversal.
+
+    Args:
+        graphql_schema: The GraphQL schema
+        root_type: The root type to start traversal from
+
+    Returns:
+        Set[GraphQLType]: Set of referenced GraphQL type objects
+    """
+    visited: set[str] = set()
+    referenced: set[GraphQLType] = set()
+
+    def visit_type(type_name: str) -> None:
+        if type_name in visited:
+            return
+
+        visited.add(type_name)
+
+        if type_name.startswith("__"):
+            return
+
+        if type_name in {"Query", "Mutation", "Subscription"}:
+            return
+
+        type_def = graphql_schema.type_map.get(type_name)
+        if not type_def:
+            return
+            
+        referenced.add(type_def)
+
+        if is_object_type(type_def) and not has_directive(cast(GraphQLObjectType, type_def), "instanceTag"):
+            visit_object_type(cast(GraphQLObjectType, type_def))
+        elif is_interface_type(type_def):
+            visit_interface_type(cast(GraphQLInterfaceType, type_def))
+        elif is_union_type(type_def):
+            visit_union_type(cast(GraphQLUnionType, type_def))
+        elif is_input_object_type(type_def):
+            visit_input_object_type(cast(GraphQLInputObjectType, type_def))
+        # Scalar and enum types don't reference other types
+
+    def visit_object_type(obj_type: GraphQLObjectType) -> None:
+        for field in obj_type.fields.values():
+            visit_field_type(field.type)
+
+        for interface in obj_type.interfaces:
+            visit_type(interface.name)
+
+    def visit_interface_type(interface_type: GraphQLInterfaceType) -> None:
+        for field in interface_type.fields.values():
+            visit_field_type(field.type)
+
+    def visit_union_type(union_type: GraphQLUnionType) -> None:
+        for member_type in union_type.types:
+            visit_type(member_type.name)
+
+    def visit_input_object_type(input_type: GraphQLInputObjectType) -> None:
+        for field in input_type.fields.values():
+            visit_field_type(field.type)
+
+    def visit_field_type(field_type: GraphQLType) -> None:
+        unwrapped_type = field_type
+        while is_non_null_type(unwrapped_type) or is_list_type(unwrapped_type):
+            if is_non_null_type(unwrapped_type):
+                unwrapped_type = cast(GraphQLNonNull[Any], unwrapped_type).of_type
+            elif is_list_type(unwrapped_type):
+                unwrapped_type = cast(GraphQLList[Any], unwrapped_type).of_type
+
+        if hasattr(unwrapped_type, "name"):
+            visit_type(unwrapped_type.name)
+
+    visit_type(root_type)
+
+    log.info(f"Found {len(referenced)} referenced types from root type '{root_type}'")
+    return referenced
